move enums into models folder
