+++ assistant/tests/models/test_reviewer.py
@@ -23,11 +23,12 @@
 #    see http://www.gnu.org/licenses/.
 #
 ##############################################################################
-import factory
 from django.test import TestCase
-from assistant.tests.factories.reviewer import ReviewerFactory
-from assistant.enums import reviewer_role
+
 from assistant.models import reviewer
+from assistant.models.enums import reviewer_role
+from assistant.tests.factories.reviewer import ReviewerFactory
+
 
 class TestReviewerFactory(TestCase):
